requirements.txt
Flask==3.0.3
Werkzeug==3.0.6
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
kubernetes==31.0.0
PyYAML==6.0.2
pydantic==2.9.2
requests==2.32.3
urllib3==2.2.3
openai==1.52.2
python-dotenv==1.0.0

